% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awatea-mcmc.R
\name{awatea_mcmc}
\alias{awatea_mcmc}
\title{Convert Awatea mcmc output into dataframe of annual medians}
\usage{
awatea_mcmc(csv, species, stock, var_name = "biomass")
}
\arguments{
\item{csv}{csv file location and name}

\item{species}{species label string}

\item{stock}{stock label string}

\item{var_name}{what to call parameter output}
}
\description{
Convert Awatea mcmc output into dataframe of annual medians
}
\examples{
d <- awatea_mcmc(
  csv = "data/BOR-CST-2021-Awatea-output-forRobyn-230119/BOR-CST-2021-MCMC(B)-forRobyn.csv",
  species = "bocaccio",
  stock = "Coastwide",
  var_name = "biomass"
)

}
