% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-prod.R
\name{calc_prod}
\alias{calc_prod}
\title{Calculate stock productivity, recruitment, and lag variables from mcmc outputs}
\usage{
calc_prod(
  catchfile,
  mcmcprefix,
  mcmcsuffix,
  species,
  stock,
  model_type = "awatea",
  start_year = NULL,
  end_year = NULL,
  recruitment_age = 1
)
}
\arguments{
\item{catchfile}{location of catch data csv summarized by year and summed across fleets}

\item{mcmcprefix}{mcmc csv file location and name prefix (everything before the variable label)}

\item{mcmcsuffix}{mcmc csv file location and name suffix (everything after the variable label)}

\item{species}{species label string}

\item{stock}{stock label string}

\item{model_type}{which stock assessment platform was used to produce mcmc outputs}

\item{start_year}{default \code{NULL}; integer value required for iscam outputs}

\item{end_year}{default \code{NULL}; integer value required for iscam outputs}

\item{recruitment_age}{default \code{1}; integer value}
}
\value{
dataframe
}
\description{
Calculate stock productivity, recruitment, and lag variables from mcmc outputs
}
\examples{
calc_prod(
  catchfile = "data/arrowtooth-mcmc/arrowtooth-total-catch.csv",
  mcmcprefix = "data/arrowtooth-mcmc/iscam_",
  mcmcsuffix = "_mcmc.csv",
  species = "arrowtooth flounder",
  stock = "Coastwide",
  model_type = "iscam",
  start_year = 1996,
  end_year = 2021,
  recruitment_age = 1
)

}
