% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cor.R
\name{cor_func}
\alias{cor_func}
\title{Plot correlations
hack found on stackoverflow https://stackoverflow.com/questions/37889222/change-colors-in-ggpairs-now-that-params-is-deprecated}
\usage{
cor_func(data, mapping, method, symbol, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot correlations
hack found on stackoverflow https://stackoverflow.com/questions/37889222/change-colors-in-ggpairs-now-that-params-is-deprecated
}
