% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm.R
\name{do_fit}
\alias{do_fit}
\title{brms helper functions: fit once per response draw}
\usage{
do_fit(dat, poly = TRUE)
}
\arguments{
\item{poly}{}
}
\description{
brms helper functions: fit once per response draw
}
