% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal-averages.R
\name{extract_enviro_var}
\alias{extract_enviro_var}
\title{Extract spatial and temporal averaged environmental variables}
\usage{
extract_enviro_var(
  variable,
  description,
  months = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12),
  grid = NULL
)
}
\arguments{
\item{variable}{Annual index (column named anamoly or value) or spatial layer from pacea}

\item{description}{Label for output variable in quotations}

\item{months}{Vector of numeric months to average}

\item{grid}{Grid with sf geometry}
}
\description{
Extract spatial and temporal averaged environmental variables
}
