% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-netcdf-values.R
\name{extract_netcdf_values}
\alias{extract_netcdf_values}
\title{Get annual averages from certain times within the netcdf}
\usage{
extract_netcdf_values(
  nc.file,
  output.dir = "data",
  xvar_name = "longitude",
  yvar_name = "latitude",
  time_name = "time",
  time_resolution_input = "monthly",
  time_resolution_output = "annual",
  variable_name = "tob",
  model_start_time = 1950,
  model_end_time = 1979,
  xbounds = c(-133.85254, -123.21362),
  ybounds = c(48.21968, 55.6862),
  whichtimes = c(1:12),
  agg_method = "mean"
)
}
\arguments{
\item{nc.file}{Path and file name for netcdf}

\item{output.dir}{Directory to save output into}

\item{xvar_name}{What is the x variable called when stored as an array, not a vector? Default "longitude"}

\item{yvar_name}{What is the y variable called when stored as an array, not a vector? Default "latitude"}

\item{time_name}{What is the time variable called? Default "time"}

\item{time_resolution_input}{Currently only working for monthly data}

\item{time_resolution_output}{Currently only working for annual values}

\item{variable_name}{What is the variable of interest called? Default "tob"}

\item{model_start_time}{Starting time in the output units. Currently on working for years}

\item{model_end_time}{Ending time in the output units. Currently on working for years}

\item{xbounds}{A vector of length 2 containing min and max values in same units as netcdf dimensions.
Default is longitude for all BC coast}

\item{ybounds}{A vector of length 2 containing min and max values in same units as netcdf dimensions.
Default is latitude for all BC coast}

\item{whichtimes}{A vector of integers representing times of interest. Default is \code{c(1:12)} for annual mean from monthly data}

\item{agg_method}{Method for aggregating data. Default "mean", other options include "min" and "max"}
}
\value{
A data frame with columns X, Y, and for each time step output.
}
\description{
Get annual averages from certain times within the netcdf
}
\examples{
df <- extract_netcdf_values(filename,
  variable_name = "tob",
  model_start_year = 1950,
  model_end_year = 1979,
  agg_method = "max")

}
