% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_months}
\alias{filter_months}
\title{Get environmental means for a specific set of months and standardize}
\usage{
filter_months(data, months, type_label)
}
\arguments{
\item{data}{Any monthly environmental index with numeric month and year columns}

\item{months}{Which months to include as an average}

\item{type_label}{}
}
\description{
Get environmental means for a specific set of months and standardize
}
