% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-ss3-summary.R
\name{format_ss3_summary}
\alias{format_ss3_summary}
\title{format SS3 summary output}
\usage{
format_ss3_summary(output, species, stock)
}
\arguments{
\item{output}{Summary output from SS3}

\item{stock}{}
}
\description{
format SS3 summary output
}
