% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stock-enviro-var.R
\name{get_stock_enviro_var}
\alias{get_stock_enviro_var}
\title{Get stock specific environmental variables from any larger scale spatiotemporal grid}
\usage{
get_stock_enviro_var(
  temporal_grid = grid,
  variable_name = "ann_mean",
  species = "NA",
  stock = "NA",
  stage = "NA",
  time_var = "year",
  recruitment_age = 1,
  depth_range = c(0, 200),
  lon_range = NULL,
  lat_range = NULL,
  lon_var_name = "UTM.lon",
  lat_var_name = "UTM.lat",
  polygon = NULL,
  grid_crs = "+proj=utm +zone=9 +datum=WGS84 +units=km +no_defs",
  polygon_crs = "+proj=utm +zone=9 +datum=WGS84 +units=km",
  bbox = NULL
)
}
\arguments{
\item{temporal_grid}{Spatiotemporal grid spanning the time frame and spatial extents relevant to a stock}

\item{variable_name}{Which variable to extract and summarize}

\item{species}{Optional argument to provide a variable indicating the species. Default is "NA"}

\item{stock}{Optional argument to provide a variable indicating the stock. Default is "NA"}

\item{stage}{Optional argument to provide a variable indicating the relevant life history stage. Default is "NA"}

\item{time_var}{Which variable contains the time steps of interest. Must be evenly spaced integers}

\item{recruitment_age}{The age at recruitment used in the stock assessments in same units as time_var
This controls which lags (up to age + 1) are included.}

\item{depth_range}{Vector length 2 defining the range of depths of interest.
Can/should depend on the variable, species, and stock. Can be \code{NULL} when all depths should be aggregated.}

\item{lon_range}{Vector length 2 defining the longitude range. Can be \code{NULL} when all should be aggregated.}

\item{lat_range}{Vector length 2 defining the latitude range. Can be \code{NULL} when all should be aggregated.}

\item{lon_var_name}{Provide name of longitude variable in the grid, if not using a spatial object}

\item{lat_var_name}{Provide name of latitude variable in the grid, if not using a spatial object}

\item{polygon}{Polygons are not yet supported}

\item{bbox}{A bbox is not yet supported}
}
\value{
A data frame of time steps and variable values summarized for a specific geographic area and depth range.
}
\description{
Get stock specific environmental variables from any larger scale spatiotemporal grid
}
\examples{
rds.file <- "data/annual_mean_tob_bc.rds"
load(file = "data/wcvi_grid.rda")

grid <- project_netcdf_values(
  rds.file,
  variable_name = "ann_mean",
  print_test_plot = TRUE,
  coord_df = wcvi_grid,
  grid_xvar_name = "X",
  grid_yvar_name = "Y"
)

df <- get_stock_enviro_var(
  temporal_grid = grid,
  variable_name = "ann_mean",
  species = "yelloweye rockfish",
  stock = "WCVI",
  recruitment_age = 1,
  depth_range = c(0, 200),
  lon_range = NULL,
  lat_range = c(48, 51.2),
  lon_var_name = "LONGITUDE",
  lat_var_name = "LATITUDE"
)

}
