% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stock-enviro-vars.R
\name{get_stock_enviro_vars}
\alias{get_stock_enviro_vars}
\title{Get stock specific environmental variables from any larger scale spatiotemporal grid}
\usage{
get_stock_enviro_vars(
  temporal_grid = grid,
  variable_name = "ann_mean",
  species = "NA",
  stock = "NA",
  recruitment_age = 1,
  depth_range = c(0, 200),
  lon_range = NULL,
  lat_range = NULL,
  lon_var_name = "UTM.lon",
  lat_var_name = "UTM.lat",
  polygon = NULL,
  bbox = NULL
)
}
\arguments{
\item{temporal_grid}{Spatiotemporal grid spanning the time frame and spatial extents relevant to a stock}

\item{variable_name}{Which variable to extract and summarize}

\item{species}{Optional argument to provide a variable indicating the species. Default is "NA"}

\item{stock}{Optional argument to provide a variable indicating the stock. Default is "NA"}

\item{recruitment_age}{The age at recruitment used in the stock assessments.
This controls which lags (up to age + 1) are included.}

\item{depth_range}{Vector length 2 defining the range of depths of interest.
Can/should depend on the variable, species, and stock. Can be \code{NULL} when all depths should be aggregated.}

\item{lon_range}{Vector length 2 defining the longitude range. Can be \code{NULL} when all should be aggregated.}

\item{lat_range}{Vector length 2 defining the latitude range. Can be \code{NULL} when all should be aggregated.}

\item{lon_var_name}{Provide name of longitude variable in the grid, if not using a spatial object}

\item{lat_var_name}{Provide name of latitude variable in the grid, if not using a spatial object}

\item{polygon}{Polygons are not yet supported}

\item{bbox}{A bbox is not yet supported}
}
\value{
A data frame of time steps and variable values summarized for a specific geographic area and depth range.
}
\description{
Get stock specific environmental variables from any larger scale spatiotemporal grid
}
\examples{
df <- get_stock_enviro_vars(  temporal_grid = grid,
variable_name = "ann_mean",
species = "yelloweye rockfish",
stock = "WCVI",
recruitment_age = 1,
depth_range = c(0, 200),
lon_range = NULL,
lat_range = c(48,51.2),
lon_var_name = "LONGITUDE",
lat_var_name = "LATITUDE"
)

}
