% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam.R
\name{iscam_mcmc}
\alias{iscam_mcmc}
\title{Convert iscam mcmc output into dataframe of annual medians}
\usage{
iscam_mcmc(csv, species, stock, start_year, end_year, var_name = "biomass")
}
\arguments{
\item{csv}{csv file location and name}

\item{species}{species label string}

\item{stock}{stock label string}

\item{start_year}{first year as integer}

\item{end_year}{last year as integer}

\item{var_name}{what to call parameter output}
}
\value{
dataframe of parameter values by year
}
\description{
Convert iscam mcmc output into dataframe of annual medians
}
\examples{
d <- iscam_mcmc(
  csv = "data/PCOD/iscam_sbt_mcmc_HS_QCS.csv",
  species = "pacific cod",
  stock = "SYN HS & QCS",
  start_year = 1956,
  end_year = 2021,
  var_name = "biomass"
)

}
