% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-netcdf-values.R
\name{project_netcdf_values}
\alias{project_netcdf_values}
\title{Get aggregated netcdf values for specific locations or grid cells}
\usage{
project_netcdf_values(
  rds.file,
  variable_name,
  set_resolution = 10000,
  print_test_plot = FALSE,
  var_xvar_name = "X",
  var_yvar_name = "Y",
  var_crs = "+proj=longlat +datum=WGS84",
  spatial_grid = bccoast,
  grid_xvar_name = "UTM.lon",
  grid_yvar_name = "UTM.lat",
  grid_crs = "+proj=utm +zone=9 +datum=WGS84"
)
}
\arguments{
\item{rds.file}{File containing output from \code{extract-netcdf()} function}

\item{variable_name}{What variable does the above file contain}

\item{set_resolution}{Choose a resolution in m for the projected raster}

\item{print_test_plot}{If \code{TRUE} will return a plot of how the resolution looks}

\item{var_xvar_name}{What is the x variable from the rds called?}

\item{var_yvar_name}{What is the Y variable from the rds called?}

\item{var_crs}{What CRS projection was the netcdf file in?}

\item{spatial_grid}{The grid to extract values for. Can be spatial object, or have x and y columns.}

\item{grid_xvar_name}{If a spatial object, use \code{NULL}, otherwise give name of x variable in the grid.}

\item{grid_yvar_name}{If a spatial object, use \code{NULL}, otherwise give name of y variable in the grid.}

\item{grid_crs}{What CRS projection is the grid in?}
}
\value{
The grid expanded to include all years and the netcdf derived variables.
}
\description{
Get aggregated netcdf values for specific locations or grid cells
}
\examples{
project_netcdf_values(rds.file,
  variable_name = "ann_mean",
  print_test_plot = TRUE,
  spatial_grid = bccoast,
  grid_xvar_name = "UTM.lon",
  grid_yvar_name = "UTM.lat"
)

}
