% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_model}
\alias{refine_model}
\title{helper functions
refine delta model}
\usage{
refine_model(m, alternate_family = set_family2, use_priors = sdmTMBpriors())
}
\description{
helper functions
refine delta model
}
