% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal-averages.R
\name{spatial_average}
\alias{spatial_average}
\title{Calculate spatial average from pacea objects and defined by these grids;
default area is roughly whole coast}
\usage{
spatial_average(
  pacea_st_obj,
  area = list(matrix(c(-134, -131, -128, -124, -125.5, -134, -134, 54.4, 54.4, 50.5, 48,
    47.5, 52, 54.4), ncol = 2))
)
}
\arguments{
\item{area}{Provide a geometry object or defaults to the whole BC coast}
}
\description{
Calculate spatial average from pacea objects and defined by these grids;
default area is roughly whole coast
}
