% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-netcdf-values.R
\name{project_netcdf_values}
\alias{project_netcdf_values}
\title{Get aggregated netcdf values for specific locations or grid cells}
\usage{
project_netcdf_values(
  nc.rds.file,
  variable_name,
  print_test_plot = FALSE,
  var_xvar_name = "X",
  var_yvar_name = "Y",
  var_crs = "+proj=longlat +datum=WGS84",
  coord_df = bccoast,
  grid_xvar_name = "UTM.lon",
  grid_yvar_name = "UTM.lat",
  grid_crs = "+proj=utm +zone=9 +datum=WGS84 +units=km",
  set_resolution = 10,
  radius = NULL
)
}
\arguments{
\item{variable_name}{What variable does the above file contain}

\item{print_test_plot}{If \code{TRUE} will return a plot of the first time step, showing how the resolution and extraction is working}

\item{var_xvar_name}{What is the x variable from the rds called? Default matches output from \code{extract_netcdf_values()} function}

\item{var_yvar_name}{What is the Y variable from the rds called? Default matches output from \code{extract_netcdf_values()} function}

\item{var_crs}{What CRS projection was the netcdf file in?}

\item{coord_df}{Coordinates data frame. Can be spatial object, or any data frame with x and y columns}

\item{grid_xvar_name}{Provide name of x variable in the grid, if not using a spatial object}

\item{grid_yvar_name}{Provide name of y variable in the grid, if not using a spatial object}

\item{grid_crs}{What CRS projection is the grid in?}

\item{set_resolution}{Choose a resolution in same units as grid_crs}

\item{radius}{Choose radius around points (only applies if not using a polygon). Default is half the resolution.}

\item{rds.file}{File containing output from \code{extract_netcdf_values()} function}
}
\value{
The grid expanded to include all years and the netcdf derived variables.
}
\description{
Get aggregated netcdf values for specific locations or grid cells
}
\examples{
rds.file <- "data/annual_mean_tob_bc.rds"
project_netcdf_values(
  rds.file,
  variable_name = "ann_mean",
  print_test_plot = TRUE,
  coord_df = sdmTMB::qcs_grid,
  grid_xvar_name = "X",
  grid_yvar_name = "Y"
)
}
